modulo_04/extracao_twitter/requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter-client==7.4.9
jupyter-core==5.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.7.3
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
tweepy==4.12.1
urllib3==1.26.14
wcwidth==0.2.6
